+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Things change, people change, everything changes.
 
-## Unreleased
+## [0.6.2](https://github.com/elfacht/craft-deploy/compare/0.6.1...0.6.2) - 2019-07-22
 ### Added
 - Added `DEPLOY_KEEP_RELEASES` constant.
 - Added `DEPLOY_BRANCH` constant to clone specific branch.